Refactored database and database tests

+++ app/src/main/java/com/klinker/android/twitter_l/data/roomdb/entities/ScheduledTweet.kt
@@ -7,10 +7,12 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+
+//TODO: decide whether to put an index on date (may not be worth it)
 @Entity(tableName = "scheduled_tweets")
 class ScheduledTweet(@field:PrimaryKey
                      @field:ColumnInfo(name = "alarm_id")
-                     var alarmId: Int, @field:ColumnInfo
+                     var alarmId: Long, @field:ColumnInfo
                      var text: String, @field:ColumnInfo
                      var time: Long, @field:ColumnInfo
                      var account: Int)
